new domain

+++ src/main/java/com/shows/as/domain/adapters/IBankServiceAdapter.java
@@ -4,6 +4,6 @@ import java.util.Date;
 
 public interface IBankServiceAdapter {
 
-    Boolean autoritza(String dni, Integer codiB, String numCompte, Float preu, Integer codiBancShows, String numCompteShows, Date dAvui);
+    Boolean autoritza(String dni, Integer codiB, String numCompte, Double preu, Integer codiBancShows, String numCompteShows, Date dAvui);
 
 }
